STOP if first u or v < -999
This suggests the wind layer is missing. I think its actually -9999.

When the wind layer is requested in the namelist, and it is not found,
while it is "NOTE"ed in the log, the program continues, treating the
-9999 u and v as valid wind speeds. These highly negative winds are
used to advect the vortex to its expected location at the next forecast
time.
The advected location is inevitably too far away from the actual new
location of the vortex. This results in no coherent track--just a lot
of individual points.